Update patch set 13
Patch Set 13:

I am just wondering; would it be beneficial for the team to discuss and document how/where we configure 'properties' (including environment variables etc). I often am unsure about this and cannot really review code related to that because the overal picture is unclear for me. I understand there a many options and maybe we are inconsistent too. Us usual there might not be agreed best approach and ONAP might also push us toward how it was usually done'

But as usual I think we need consensus as a team for the CPS project and I fell it warrant a quick spiek for someone to look into this and document our preferred way for dealing with (build) properties and how to override them etc.

But maybe it just me and everyone els eis clear on this and we don't need this, lets discuss in todays standup.

Patch-set: 13
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>