Update patch set 10
Patch Set 10:

(2 comments)

HI Bruno, we started a bit of a team review on this (didn't get far) but I have recorded 2 questions on why you made certain changes. 

I know we had many discussion on the encv variables but it was a bit hard to follow for me and maybe the 2nd question is a good real-world example of what those discussion mean for our code.

Maybe we can take this in the standup today?

Patch-set: 10
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>